+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/SavepointNode.java
@@ -63,7 +63,7 @@ public class SavepointNode extends DDLStatementNode
 	{
 		initAndCheck(null);	
 		this.savepointName = (String) objectName;
-		this.savepointStatementType = ((Integer) savepointStatementType).intValue();
+		this.savepointStatementType = (Integer) savepointStatementType;
 
 		if (SanityManager.DEBUG)
 		{
